fixed crash with telekinetic grab (ExpPD)

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/spells/TelekineticGrab.java
@@ -95,25 +95,29 @@ public class TelekineticGrab extends TargetedSpell {
 
 		} else if (Dungeon.level.heaps.get(bolt.collisionPos) != null){
 
-			Heap h = Dungeon.level.heaps.get(bolt.collisionPos);
+			for (int n : PathFinder.NEIGHBOURS9){
+				int cell = bolt.collisionPos + n;
 
-			if (h.type != Heap.Type.HEAP){
-				GLog.w(Messages.get(this, "cant_grab"));
-				h.sprite.drop();
-				return;
-			}
+				Heap h = Dungeon.level.heaps.get(cell);
+				if (h != null) {
 
-			while (!h.isEmpty()) {
-				Item item = h.peek();
-				if (item.doPickUp(hero, h.pos)) {
-					h.pickUp();
-					hero.spend(-Item.TIME_TO_PICK_UP); //casting the spell already takes a turn
-					GLog.i( Messages.capitalize(Messages.get(hero, "you_now_have", item.name())) );
+					if (h.type != Heap.Type.HEAP) {
+						GLog.w(Messages.get(this, "cant_grab"));
+						h.sprite.drop();
+						continue;
+					}
 
-				} else {
-					GLog.w(Messages.get(this, "cant_grab"));
-					h.sprite.drop();
-					return;
+					while (!h.isEmpty()) {
+						Item item = h.peek();
+						if (item.doPickUp(hero, h.pos)) {
+							h.pickUp();
+							hero.spend(-Item.TIME_TO_PICK_UP); //casting the spell already takes a turn
+							GLog.i(Messages.capitalize(Messages.get(hero, "you_now_have", item.name())));
+						} else {
+							GLog.w(Messages.get(this, "cant_grab"));
+							h.sprite.drop();
+						}
+					}
 				}
 			}
